alteracao do atributo potencia para bigdecimal

+++ src/main/java/com/cafe/modelo/Maquina.java
@@ -54,7 +54,7 @@ public class Maquina {
 	private BigDecimal valor;
 	
 	@PositiveOrZero
-	private Long potencia;
+	private BigDecimal potencia;
 	
 	@PositiveOrZero
 	private Integer vidaUtil;
